initial updates for refactor
